+++ src/binding/mod_keys.rs
@@ -26,13 +26,29 @@ pub struct ModKeys(u8);
 bitflags! {
     impl ModKeys: u8 {
         /// Corresponds to [`KeyCode::ControlLeft`] and [`KeyCode::ControlRight`].
-        const CONTROL = 0b00000001;
-        /// Corresponds to [`KeyCode::ShiftLeft`] and [`KeyCode::ShiftRight`]
-        const SHIFT = 0b00000010;
+        const CONTROL = 0b00010001;
+        /// Corresponds to [`KeyCode::ControlLeft`].
+        const CONTROL_LEFT = 0b00000001;
+        /// Corresponds to [`KeyCode::ControlRight`].
+        const CONTROL_RIGHT = 0b00010000;
+        /// Corresponds to [`KeyCode::ShiftLeft`] and [`KeyCode::ShiftRight`].
+        const SHIFT = 0b00100010;
+        /// Corresponds to [`KeyCode::ShiftLeft`].
+        const SHIFT_LEFT = 0b00000010;
+        /// Corresponds to [`KeyCode::ShiftRight`].
+        const SHIFT_RIGHT = 0b00100000;
         /// Corresponds to [`KeyCode::AltLeft`] and [`KeyCode::AltRight`].
-        const ALT = 0b00000100;
+        const ALT = 0b01000100;
+        /// Corresponds to [`KeyCode::AltLeft`].
+        const ALT_LEFT = 0b00000100;
+        /// Corresponds to [`KeyCode::AltRight`].
+        const ALT_RIGHT = 0b01000000;
         /// Corresponds to [`KeyCode::SuperLeft`] and [`KeyCode::SuperRight`].
-        const SUPER = 0b00001000;
+        const SUPER = 0b10001000;
+        /// Corresponds to [`KeyCode::SuperLeft`].
+        const SUPER_LEFT = 0b00001000;
+        /// Corresponds to [`KeyCode::SuperRight`].
+        const SUPER_RIGHT = 0b10000000;
     }
 }
 
@@ -58,9 +74,17 @@ impl Display for ModKeys {
             }
             match mod_key {
                 ModKeys::CONTROL => write!(f, "Ctrl")?,
+                ModKeys::CONTROL_LEFT => write!(f, "LeftCtrl")?,
+                ModKeys::CONTROL_RIGHT => write!(f, "RightCtrl")?,
                 ModKeys::SHIFT => write!(f, "Shift")?,
+                ModKeys::SHIFT_LEFT => write!(f, "LeftShift")?,
+                ModKeys::SHIFT_RIGHT => write!(f, "RightShift")?,
                 ModKeys::ALT => write!(f, "Alt")?,
+                ModKeys::ALT_LEFT => write!(f, "LeftAlt")?,
+                ModKeys::ALT_RIGHT => write!(f, "RightAlt")?,
                 ModKeys::SUPER => write!(f, "Super")?,
+                ModKeys::SUPER_LEFT => write!(f, "LeftSuper")?,
+                ModKeys::SUPER_RIGHT => write!(f, "RightSuper")?,
                 _ => unreachable!("iteration should yield only named flags"),
             }
         }
@@ -72,26 +96,48 @@ impl Display for ModKeys {
 impl ModKeys {
     /// Returns an instance with currently active modifiers.
     #[must_use]
-    pub fn pressed(keys: &ButtonInput<KeyCode>) -> Self {
+    pub fn pressed(keys: &ButtonInput<KeyCode>, unique_side: bool) -> Self {
         let mut mod_keys = Self::empty();
-        for [key1, key2] in Self::all().iter_keys() {
-            if keys.any_pressed([key1, key2]) {
-                mod_keys |= key1.into();
-            }
+        for modifier in Self::all().iter_keys() {
+            modifier.into_iter().for_each(|key: KeyCode| {
+                if keys.pressed(key) {
+                    let mod_key: ModKeys = key.into();
+                    mod_keys |= mod_key;
+                    if !unique_side {
+                        if ModKeys::CONTROL.intersects(mod_key) {
+                            mod_keys |= ModKeys::CONTROL;
+                        } else if ModKeys::SHIFT.intersects(mod_key) {
+                            mod_keys |= ModKeys::SHIFT;
+                        } else if ModKeys::ALT.intersects(mod_key) {
+                            mod_keys |= ModKeys::ALT;
+                        } else if ModKeys::SUPER.intersects(mod_key) {
+                            mod_keys |= ModKeys::SUPER;
+                        }
+                    } else {
+                        mod_keys |= mod_key;
+                    }
+                }
+            });
         }
 
         mod_keys
     }
 
     /// Returns an iterator over the key codes corresponding to the set modifier bits.
-    ///
-    /// Each item contains left and right key codes.
-    pub fn iter_keys(self) -> impl Iterator<Item = [KeyCode; 2]> {
+    pub fn iter_keys(self) -> impl Iterator<Item = Vec<KeyCode>> {
         self.iter_names().map(|(_, mod_key)| match mod_key {
-            ModKeys::CONTROL => [KeyCode::ControlLeft, KeyCode::ControlRight],
-            ModKeys::SHIFT => [KeyCode::ShiftLeft, KeyCode::ShiftRight],
-            ModKeys::ALT => [KeyCode::AltLeft, KeyCode::AltRight],
-            ModKeys::SUPER => [KeyCode::SuperLeft, KeyCode::SuperRight],
+            ModKeys::CONTROL => vec![KeyCode::ControlLeft, KeyCode::ControlRight],
+            ModKeys::CONTROL_LEFT => vec![KeyCode::ControlLeft],
+            ModKeys::CONTROL_RIGHT => vec![KeyCode::ControlRight],
+            ModKeys::SHIFT => vec![KeyCode::ShiftLeft, KeyCode::ShiftRight],
+            ModKeys::SHIFT_LEFT => vec![KeyCode::ShiftLeft],
+            ModKeys::SHIFT_RIGHT => vec![KeyCode::ShiftRight],
+            ModKeys::ALT => vec![KeyCode::AltLeft, KeyCode::AltRight],
+            ModKeys::ALT_LEFT => vec![KeyCode::AltLeft],
+            ModKeys::ALT_RIGHT => vec![KeyCode::AltRight],
+            ModKeys::SUPER => vec![KeyCode::SuperLeft, KeyCode::SuperRight],
+            ModKeys::SUPER_LEFT => vec![KeyCode::SuperLeft],
+            ModKeys::SUPER_RIGHT => vec![KeyCode::SuperRight],
             _ => unreachable!("iteration should yield only named flags"),
         })
     }
@@ -103,10 +149,14 @@ impl From<KeyCode> for ModKeys {
     /// Returns [`ModKeys::empty`] if the key is not a modifier.
     fn from(value: KeyCode) -> Self {
         match value {
-            KeyCode::ControlLeft | KeyCode::ControlRight => ModKeys::CONTROL,
-            KeyCode::ShiftLeft | KeyCode::ShiftRight => ModKeys::SHIFT,
-            KeyCode::AltLeft | KeyCode::AltRight => ModKeys::ALT,
-            KeyCode::SuperLeft | KeyCode::SuperRight => ModKeys::SUPER,
+            KeyCode::ControlLeft => ModKeys::CONTROL_LEFT,
+            KeyCode::ControlRight => ModKeys::CONTROL_RIGHT,
+            KeyCode::ShiftLeft => ModKeys::SHIFT_LEFT,
+            KeyCode::ShiftRight => ModKeys::SHIFT_RIGHT,
+            KeyCode::AltLeft => ModKeys::ALT_LEFT,
+            KeyCode::AltRight => ModKeys::ALT_RIGHT,
+            KeyCode::SuperLeft => ModKeys::SUPER_LEFT,
+            KeyCode::SuperRight => ModKeys::SUPER_RIGHT,
             _ => ModKeys::empty(),
         }
     }
@@ -125,8 +175,23 @@ mod tests {
         keys.press(KeyCode::ShiftLeft);
         keys.press(KeyCode::KeyC);
 
-        let mod_keys = ModKeys::pressed(&keys);
+        // Checking for either key
+        let mod_keys = ModKeys::pressed(&keys, false);
+        assert!(mod_keys.intersects(ModKeys::CONTROL));
+        assert!(mod_keys.intersects(ModKeys::SHIFT));
         assert_eq!(mod_keys, ModKeys::CONTROL | ModKeys::SHIFT);
+        // This is currently not equal to either key
+        assert_ne!(mod_keys, ModKeys::CONTROL_LEFT | ModKeys::SHIFT_LEFT);
+
+        // Checking for unique key
+        let mod_keys = ModKeys::pressed(&keys, true);
+        assert_eq!(mod_keys, ModKeys::CONTROL_LEFT | ModKeys::SHIFT_LEFT);
+        assert_ne!(mod_keys, ModKeys::CONTROL_RIGHT | ModKeys::SHIFT_RIGHT);
+        // This is currently not equal to the unique keys
+        assert_ne!(mod_keys, ModKeys::CONTROL | ModKeys::SHIFT);
+        // Even though it intersects with them
+        assert!(mod_keys.intersects(ModKeys::CONTROL));
+        assert!(mod_keys.intersects(ModKeys::SHIFT));
     }
 
     #[test]
@@ -134,6 +199,7 @@ mod tests {
         assert_eq!(ModKeys::CONTROL.to_string(), "Ctrl");
         assert_eq!(ModKeys::all().to_string(), "Ctrl + Shift + Alt + Super");
         assert_eq!(ModKeys::empty().to_string(), "");
+        assert_eq!(ModKeys::ALT_LEFT.to_string(), "LeftAlt");
     }
 
     #[cfg(feature = "serialize")]
@@ -145,8 +211,15 @@ mod tests {
             "\"CONTROL | SHIFT\""
         );
         assert_eq!(ron::to_string(&ModKeys::empty()).unwrap(), "\"\"");
+        assert_eq!(
+            ron::to_string(&ModKeys::SUPER_LEFT).unwrap(),
+            "\"LEFT_SUPER\""
+        );
 
         let parsed: ModKeys = ron::from_str("\"CONTROL | SHIFT\"").unwrap();
         assert_eq!(parsed, ModKeys::CONTROL | ModKeys::SHIFT);
+
+        let parsed: ModKeys = ron::from_str("\"RIGHT_ALT | RIGHT_SUPER\"").unwrap();
+        assert_eq!(parsed, ModKeys::ALT_RIGHT | ModKeys::SUPER_RIGHT);
     }
 }